re-org macros
